refactor: 의미없는 scope function 제거, LottoOrganizer의 책임 LottoPerson으로 이동
