+++ src/main/kotlin/lotto/LottoOrganizer.kt
@@ -1,9 +1,6 @@
 package lotto
 
-import lotto.collection.LottoNumber
-import lotto.collection.LottoResults
-import lotto.collection.LottoTicket
-import lotto.model.LottoGenerator
+import lotto.model.*
 import lotto.view.InputView
 import lotto.view.ResultView
 
@@ -11,21 +8,18 @@ import lotto.view.ResultView
 class LottoOrganizer {
 
     fun start() {
-        getTicketCount().let { ResultView.renderTicketCount(it) }
-            .let { LottoGenerator.generateTickets(it) { this.shuffled() } }
-            .let { ResultView.renderTickets(it) }.let(validateTickets).let {
-                ResultView.renderResults(
-                    it
-                )
-            }
+        val ticketCount = LottoPerson.buyLottoTicket(InputView.getPurchasePrice())
+
+        ResultView.renderTicketCount(ticketCount)
+
+        val tickets = LottoGenerator.generateTickets(ticketCount) { this.shuffled() }
+
+        ResultView.renderTickets(tickets)
+
+        val results = LottoResults(tickets, InputView.getWinningNumbers().map { LottoNumber.from(it) })
+
+        ResultView.renderResults(
+            results
+        )
     }
-
-    private fun getTicketCount(): Int = InputView.getPurchasePrice() / LottoTicket.TICKET_PRICE
-
-    private val validateTickets = { lottoTickets: List<LottoTicket> ->
-        LottoResults(
-            lottoTickets,
-            InputView.getWinningNumbers().map { LottoNumber.from(it) })
-    }
-
 }
