+++ src/main/kotlin/lotto/model/LottoPerson.kt
@@ -1,0 +1,9 @@
+package lotto.model
+
+class LottoPerson {
+
+    companion object{
+        fun buyLottoTicket(purchasePrice: Int): Int = purchasePrice / LottoTicket.TICKET_PRICE;
+
+    }
+}
