+++ src/test/kotlin/lotto/model/ProfitCalculatorTest.kt
@@ -2,7 +2,6 @@ package lotto.model
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import lotto.collection.LottoResults
 
 class ProfitCalculatorTest : BehaviorSpec({
     Given(
